fix: address cli review feedback
